pmic: pca9450: Fix handling of LDO5 regulator
The LDO5 regulator is special as it has two control registers and the
register that is used for sourcing the regulators voltage depends on
the state of the SD_VSEL input.

To correctly retrieve the actual voltage of the LD05 output, we need
to select the correct register according to the current state of
SD_VSEL.

For setting the voltage level of LD05 we assume that SD_VSEL is
controlled independently (usually by setting/clearing VSELECT in
the ESDHC driver) and we therefore try to avoid writing values
other than the defaults to the registers.

To make this work the SION bit must be set in the IOMUX for the
GPIO1_IO04 pad, so we can use the VSELECT function of the ESDHC
controller while getting the current state of the signal via GPIO
in the regulator driver.

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>